Refactor tenet-web.rs into modular src/web_client/ structure
Break the 3437-line monolithic tenet-web.rs binary into a clean
separation of concerns across 18 focused modules:

- web_client/config.rs: CLI args, environment config, constants
- web_client/state.rs: AppState, SharedState, WsEvent types
- web_client/utils.rs: api_error, message_to_json, link_attachments
- web_client/router.rs: Axum router construction
- web_client/sync.rs: Background relay sync, envelope processing
- web_client/static_files.rs: Embedded SPA serving
- web_client/handlers/{health,messages,peers,friends,groups,
  attachments,reactions,replies,profiles,conversations,websocket}.rs

Move fundamental protocol operations to the core library:
- relay_transport.rs: post_envelope() and fetch_inbox() for any
  tenet client to communicate with relay servers
- crypto.rs: validate_hex_key() for cryptographic key validation

The binary (src/bin/tenet-web.rs) is now 9 lines, delegating to
tenet::web_client::run().

All 90 tests pass. No behavioral changes.
