+++ src/lib.rs
@@ -5,5 +5,7 @@ pub mod identity;
 pub mod logging;
 pub mod protocol;
 pub mod relay;
+pub mod relay_transport;
 pub mod simulation;
 pub mod storage;
+pub mod web_client;
