+++ src/web_client/handlers/mod.rs
@@ -1,0 +1,13 @@
+//! Route handler modules for the tenet-web REST API.
+
+pub mod attachments;
+pub mod conversations;
+pub mod friends;
+pub mod groups;
+pub mod health;
+pub mod messages;
+pub mod peers;
+pub mod profiles;
+pub mod reactions;
+pub mod replies;
+pub mod websocket;
