+++ src/web_client/static_files.rs
@@ -1,0 +1,38 @@
+//! Embedded static asset serving for the SPA.
+
+use axum::http::{header, StatusCode};
+use axum::response::{IntoResponse, Response};
+use rust_embed::Embed;
+
+#[derive(Embed)]
+#[folder = "web/dist/"]
+struct Assets;
+
+pub async fn static_handler(uri: axum::http::Uri) -> Response {
+    let path = uri.path().trim_start_matches('/');
+    let path = if path.is_empty() { "index.html" } else { path };
+
+    match Assets::get(path) {
+        Some(content) => {
+            let mime = mime_guess::from_path(path).first_or_octet_stream();
+            (
+                StatusCode::OK,
+                [(header::CONTENT_TYPE, mime.as_ref())],
+                content.data.to_vec(),
+            )
+                .into_response()
+        }
+        None => {
+            // SPA fallback: serve index.html for unmatched routes
+            match Assets::get("index.html") {
+                Some(content) => (
+                    StatusCode::OK,
+                    [(header::CONTENT_TYPE, "text/html")],
+                    content.data.to_vec(),
+                )
+                    .into_response(),
+                None => (StatusCode::NOT_FOUND, "not found").into_response(),
+            }
+        }
+    }
+}
